Adding helper functions for plants.
